+++ src/components/ComboBox/useOptions.ts
@@ -1,4 +1,5 @@
 import { useCallback, useMemo } from 'react'
+import innerText from 'react-innertext'
 
 import { useId } from '../../hooks/useId'
 
@@ -72,7 +73,7 @@ export function useOptions<T>({
     }
     return allOptions.filter(({ item: { label } }) => {
       if (!inputValue) return true
-      return convertMatchableString(label).includes(convertMatchableString(inputValue))
+      return convertMatchableString(innerText(label)).includes(convertMatchableString(inputValue))
     })
   }, [allOptions, inputValue, isFilteringDisabled])